Move style files into a styles folder
