+++ PROJECTS/clone-tesla/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import '~/assets/index.css';
+import '~/assets/styles/index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
